+++ src/main/java/com/test/vp/card/Suit.java
@@ -1,4 +1,4 @@
-package com.test.vp;
+package com.test.vp.card;
 
 public enum Suit {
 	SPADES, HEARTS, DIAMONDS, CLUBS;
